:heavy_plus_sign: Add cache and user packages
