+++ Dockerfile.dev
@@ -1,0 +1,25 @@
+FROM python:3.10.8
+ARG TARGETARCH
+
+RUN curl -LSfs https://deb.nodesource.com/setup_18.x | bash \
+    && apt-get install -y nodejs
+
+RUN adduser --disabled-password --uid 1000 acronyms \
+    && mkdir /app \
+    && chown acronyms:acronyms /app
+
+USER acronyms
+WORKDIR /app
+ENV \
+    HOME=/home/acronyms \
+    PATH="/home/acronyms/.local/bin:${PATH}" \
+    POETRY_VIRTUALENVS_IN_PROJECT=true
+
+RUN curl -LSfs https://install.python-poetry.org | python3 -
+
+COPY --chown=acronyms . /app
+
+RUN npm ci && poetry install
+
+EXPOSE 8000
+ENTRYPOINT ["npm", "run", "dev"]
